Rename main to party and move tests to unit_tests directory
