+++ tests/unit_tests/test_wave_02.py
@@ -1,5 +1,5 @@
 import pytest
-from viewing_party.main import *
+from viewing_party.party import *
 
 
 def test_calculates_watched_average_rating():
