+++ tests/unit_tests/test_wave_04.py
@@ -1,5 +1,5 @@
 import pytest
-from viewing_party.main import *
+from viewing_party.party import *
 
 
 def test_get_available_friend_rec():
