+++ tests/unit_tests/test_wave_05.py
@@ -1,5 +1,5 @@
 import pytest
-from viewing_party.main import *
+from viewing_party.party import *
 
 
 def test_new_genre_rec():
